% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.R
\name{checkmate.testthat}
\alias{checkmate.testthat}
\alias{expect_access}
\alias{expect_array}
\alias{expect_atomic}
\alias{expect_atomic_vector}
\alias{expect_bit}
\alias{expect_character}
\alias{expect_choice}
\alias{expect_class}
\alias{expect_complex}
\alias{expect_count}
\alias{expect_data_frame}
\alias{expect_data_table}
\alias{expect_date}
\alias{expect_directory_exists}
\alias{expect_disjunct}
\alias{expect_double}
\alias{expect_environment}
\alias{expect_factor}
\alias{expect_file_exists}
\alias{expect_flag}
\alias{expect_formula}
\alias{expect_function}
\alias{expect_int}
\alias{expect_integer}
\alias{expect_integerish}
\alias{expect_list}
\alias{expect_logical}
\alias{expect_matrix}
\alias{expect_multi_class}
\alias{expect_named}
\alias{expect_names}
\alias{expect_number}
\alias{expect_numeric}
\alias{expect_path_for_output}
\alias{expect_posixct}
\alias{expect_r6}
\alias{expect_raw}
\alias{expect_scalar}
\alias{expect_scalar_na}
\alias{expect_set_equal}
\alias{expect_string}
\alias{expect_subset}
\alias{expect_tibble}
\title{Expectations for 'testthat'}
\usage{
expect_access(x, access = "", info = NULL,
  label = checkmate::vname(x))

expect_array(x, mode = NULL, any.missing = TRUE, d = NULL,
  min.d = NULL, max.d = NULL, null.ok = FALSE, info = NULL,
  label = checkmate::vname(x))

expect_atomic(x, any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, names = NULL,
  info = NULL, label = checkmate::vname(x))

expect_atomic_vector(x, any.missing = TRUE, all.missing = TRUE,
  len = NULL, min.len = NULL, max.len = NULL, unique = FALSE,
  names = NULL, info = NULL, label = checkmate::vname(x))

expect_bit(x, len = NULL, min.len = NULL, max.len = NULL,
  min.0 = NULL, min.1 = NULL, null.ok = FALSE, info = NULL,
  label = checkmate::vname(x))

expect_character(x, min.chars = NULL, pattern = NULL, fixed = NULL,
  ignore.case = FALSE, any.missing = TRUE, all.missing = TRUE,
  len = NULL, min.len = NULL, max.len = NULL, unique = FALSE,
  sorted = FALSE, names = NULL, null.ok = FALSE, info = NULL,
  label = checkmate::vname(x))

expect_choice(x, choices, null.ok = FALSE, fmatch = FALSE,
  info = NULL, label = checkmate::vname(x))

expect_class(x, classes, ordered = FALSE, null.ok = FALSE,
  info = NULL, label = checkmate::vname(x))

expect_complex(x, any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, names = NULL,
  null.ok = FALSE, info = NULL, label = checkmate::vname(x))

expect_count(x, na.ok = FALSE, positive = FALSE,
  tol = sqrt(.Machine$double.eps), null.ok = FALSE, info = NULL,
  label = checkmate::vname(x))

expect_data_frame(x, types = character(0L), any.missing = TRUE,
  all.missing = TRUE, min.rows = NULL, max.rows = NULL,
  min.cols = NULL, max.cols = NULL, nrows = NULL, ncols = NULL,
  row.names = NULL, col.names = NULL, null.ok = FALSE, info = NULL,
  label = checkmate::vname(x))

expect_data_table(x, key = NULL, index = NULL, types = character(0L),
  any.missing = TRUE, all.missing = TRUE, min.rows = NULL,
  max.rows = NULL, min.cols = NULL, max.cols = NULL, nrows = NULL,
  ncols = NULL, row.names = NULL, col.names = NULL,
  null.ok = FALSE, info = NULL, label = checkmate::vname(x))

expect_date(x, lower = NULL, upper = NULL, any.missing = TRUE,
  all.missing = TRUE, len = NULL, min.len = NULL, max.len = NULL,
  unique = FALSE, null.ok = FALSE, info = NULL,
  label = checkmate::vname(x))

expect_directory_exists(x, access = "", info = NULL,
  label = checkmate::vname(x))

expect_disjunct(x, y, fmatch = FALSE, info = NULL,
  label = checkmate::vname(x))

expect_double(x, lower = -Inf, upper = Inf, finite = FALSE,
  any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, sorted = FALSE,
  names = NULL, null.ok = FALSE, info = NULL,
  label = checkmate::vname(x))

expect_environment(x, contains = character(0L), null.ok = FALSE,
  info = NULL, label = checkmate::vname(x))

expect_factor(x, levels = NULL, ordered = NA, empty.levels.ok = TRUE,
  any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, n.levels = NULL,
  min.levels = NULL, max.levels = NULL, unique = FALSE,
  names = NULL, null.ok = FALSE, info = NULL,
  label = checkmate::vname(x))

expect_file_exists(x, access = "", extension = NULL, info = NULL,
  label = checkmate::vname(x))

expect_flag(x, na.ok = FALSE, null.ok = FALSE, info = NULL,
  label = checkmate::vname(x))

expect_formula(x, null.ok = FALSE, info = NULL,
  label = checkmate::vname(x))

expect_function(x, args = NULL, ordered = FALSE, nargs = NULL,
  null.ok = FALSE, info = NULL, label = checkmate::vname(x))

expect_int(x, na.ok = FALSE, lower = -Inf, upper = Inf,
  tol = sqrt(.Machine$double.eps), null.ok = FALSE, info = NULL,
  label = checkmate::vname(x))

expect_integer(x, lower = -Inf, upper = Inf, any.missing = TRUE,
  all.missing = TRUE, len = NULL, min.len = NULL, max.len = NULL,
  unique = FALSE, sorted = FALSE, names = NULL, null.ok = FALSE,
  info = NULL, label = checkmate::vname(x))

expect_integerish(x, tol = sqrt(.Machine$double.eps), lower = -Inf,
  upper = Inf, any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, sorted = FALSE,
  names = NULL, null.ok = FALSE, info = NULL,
  label = checkmate::vname(x))

expect_list(x, types = character(0L), any.missing = TRUE,
  all.missing = TRUE, len = NULL, min.len = NULL, max.len = NULL,
  unique = FALSE, names = NULL, null.ok = FALSE, info = NULL,
  label = checkmate::vname(x))

expect_logical(x, any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, names = NULL,
  null.ok = FALSE, info = NULL, label = checkmate::vname(x))

expect_matrix(x, mode = NULL, any.missing = TRUE, all.missing = TRUE,
  min.rows = NULL, max.rows = NULL, min.cols = NULL,
  max.cols = NULL, nrows = NULL, ncols = NULL, row.names = NULL,
  col.names = NULL, null.ok = FALSE, info = NULL,
  label = checkmate::vname(x))

expect_multi_class(x, classes, null.ok = FALSE, info = NULL,
  label = checkmate::vname(x))

expect_named(x, type = "named", info = NULL,
  label = checkmate::vname(x))

expect_names(x, type = "named", subset.of = NULL,
  must.include = NULL, permutation.of = NULL, identical.to = NULL,
  disjunct.from = NULL, what = "names", info = NULL,
  label = checkmate::vname(x))

expect_number(x, na.ok = FALSE, lower = -Inf, upper = Inf,
  finite = FALSE, null.ok = FALSE, info = NULL,
  label = checkmate::vname(x))

expect_numeric(x, lower = -Inf, upper = Inf, finite = FALSE,
  any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, sorted = FALSE,
  names = NULL, null.ok = FALSE, info = NULL,
  label = checkmate::vname(x))

expect_path_for_output(x, overwrite = FALSE, extension = NULL,
  info = NULL, label = checkmate::vname(x))

expect_posixct(x, lower = NULL, upper = NULL, any.missing = TRUE,
  all.missing = TRUE, len = NULL, min.len = NULL, max.len = NULL,
  unique = FALSE, sorted = FALSE, null.ok = FALSE, info = NULL,
  label = checkmate::vname(x))

expect_r6(x, classes = NULL, ordered = FALSE, cloneable = NULL,
  public = NULL, private = NULL, null.ok = FALSE, info = NULL,
  label = checkmate::vname(x))

expect_raw(x, len = NULL, min.len = NULL, max.len = NULL,
  names = NULL, null.ok = FALSE, info = NULL,
  label = checkmate::vname(x))

expect_scalar(x, na.ok = FALSE, null.ok = FALSE, info = NULL,
  label = checkmate::vname(x))

expect_scalar_na(x, null.ok = FALSE, info = NULL,
  label = checkmate::vname(x))

expect_set_equal(x, y, ordered = FALSE, fmatch = FALSE, info = NULL,
  label = checkmate::vname(x))

expect_string(x, na.ok = FALSE, min.chars = NULL, pattern = NULL,
  fixed = NULL, ignore.case = FALSE, null.ok = FALSE, info = NULL,
  label = checkmate::vname(x))

expect_subset(x, choices, empty.ok = TRUE, fmatch = FALSE,
  info = NULL, label = checkmate::vname(x))

expect_tibble(x, types = character(0L), any.missing = TRUE,
  all.missing = TRUE, min.rows = NULL, max.rows = NULL,
  min.cols = NULL, max.cols = NULL, nrows = NULL, ncols = NULL,
  row.names = NULL, col.names = NULL, null.ok = FALSE, info = NULL,
  label = checkmate::vname(x))
}
\arguments{
\item{info}{[character(1)]\cr
Extra information to be included in the message for the \CRANpkg{testthat} reporter.
See \code{\link[testthat]{expect_that}}.}

\item{label}{[\code{character(1)}]\cr
Name of the checked object to print in messages. Defaults to
the heuristic implemented in \code{\link{vname}}.}
}
\description{
Expectations for package \CRANpkg{testthat}.
See the respective help page in \CRANpkg{checkmate} (replace `expect_` with `check_`) for a
description of the function arguments.
E.g., for the help page of `expect_numeric`, see \code{link[checkmate]{check_numeric}}.
}
\keyword{internal}
